// apps/trade-pro/src/pages/dashboards/salesDataBranchWise/types.ts
export type TSalesDataSearchCriteria = {
  OrganizationId: number;
  CompanyId: number;
  FromDate: Date;
  ToDate: Date;
  PaymenetTermId: number; // Cash
  //,"PaymenetTermId": 2 // Credit
  WeightType: number;
  ReqType: string;
  ValueType: number;
  VoucherType: string;
  DateType: number;
  CreditSale: boolean;
  CashSale: boolean;
};

export type TSalesDataBranchWiseHistory = {
  ItemTypeId: number;
  ItemCategoryId: number;
  CategoryDescription: string;
  TypeDescription: string;
  ItemId: number;
  ItemName: string;
  Parms_01: string;
  Parms_02: string;
  Parms_03: string;
  Parms_04: string;
  Parms_05: string;
  Parms_06: string;
  Parms_07: string;
  Parms_08: string;
  Parms_09: string | null;
  Parms_10: string | null;
  Value_01: number;
  Value_02: number;
  Value_03: number;
  Value_04: number;
  Value_05: number;
  Value_06: number;
  Value_07: number;
  Value_08: number;
  Value_09: number;
  Value_10: number;
  ValueWT_01: number;
  ValueWT_02: number;
  ValueWT_03: number;
  ValueWT_04: number;
  ValueWT_05: number;
  ValueWT_06: number;
  ValueWT_07: number;
  ValueWT_08: number;
  ValueWT_09: number;
  ValueWT_10: number;
  Id_01: number;
  Id_02: number;
  Id_03: number;
  Id_04: number;
  Id_05: number;
  Id_06: number;
  Id_07: number;
  Id_08: number;
  Id_09: number;
  Id_10: number;
  FromDate: string;
  ToDate: string;
  PaymentTerm: string;
  WeightText: string;
  ValueText: string;
  ReportCriteria: string;
};